Update bindings for libhandy v0.0.10
This update contains an API breakage in the form of
changing the parent types of a few classes. Because the
new parent only exists in 0.0.10, the bindings can only
compile for 0.0.10. Thus, the version bump signifies that
users must use versions before 0.4.0 in order to target
older versions of libhandy.